requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# HTTP client for scraping
httpx>=0.25.0

# Web scraping (for court auction site)
selenium>=4.15.0
webdriver-manager>=4.0.0
beautifulsoup4>=4.12.0

# Video/Audio processing
ffmpeg-python>=0.2.0
pydub>=0.25.0
Pillow>=10.0.0

# LLM integration
anthropic>=0.18.0

# Web API server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Optional: for real TTS (later)
# google-cloud-texttospeech>=2.14.0

# Development
pytest>=7.0.0
pytest-asyncio>=0.21.0
